fix(notificaciones): castear $1 a int en ventana de CxP por vencer
CURRENT_DATE + $1 fallaba con 'operator is not unique: date + unknown':
el driver envia el parametro sin tipo y Postgres no puede desambiguar
entre date + integer y date + interval.

Este error quedaba oculto detras del de columna inexistente (cxp.total),
porque Postgres resuelve nombres de columna antes que operadores. Al
corregir la columna en f3e839e, este salio a la superficie y el cron
cronAlertasDiarias seguia crasheando.

Verificado contra el RDS: la query corre limpia y devuelve filas bien
formadas (COM-101, montoPendiente 102428.25, diasVence 28).

+++ hi-cloud backend-project/backend/src/notificaciones/notificaciones.service.ts
@@ -121,7 +121,7 @@ export class NotificacionesService {
        JOIN proveedores p ON p.id = cxp."proveedorId"
        WHERE cxp."isActive" = true
          AND cxp.estado IN ('pendiente','pagada_parcial')
-         AND cxp."fechaVencimiento"::date BETWEEN CURRENT_DATE AND CURRENT_DATE + $1
+         AND cxp."fechaVencimiento"::date BETWEEN CURRENT_DATE AND CURRENT_DATE + $1::int
        ORDER BY "diasVence" ASC
        LIMIT 50`,
       [diasAlerta],
